initial push for farmer-lib transition
