--- crates/subspace-farmer/src/commands.rs
@@ -1,3 +1,0 @@
-mod farm;
-
-pub(crate) use farm::farm;
